Only allocate if we have security enabled
The section of this code which allocates (and it is therefore slower) is
only needed if socket security is enabled. Otherwise, we don't care
about the remote end point in the ReceiveFrom_nochecks_exc method, so it
is better to just call Receive_internal and avoid the overhead of the
allocation for the EndPoint object.

This correct unity case 763848.