Make xbuild on Windows use Mono instead of .NET
We don't have .NEt 4.6 installed on all of our build machines, to
running mcs.exe on Windows fails, as it was built with .NET 4.6.
Instead, we will run mcs.exe via Mono on Windows (which is what happens
on OSX, although there the low-level Process class handles it).

To accomplish this, we run mcs.bat instead of mcs.exe.