Exit async io worker threads on domain unload
This change corrects case 723177 and case 759191.

The async IO worker threads are domain agnostic. That is, they connect
to a given domain to execute a managed method on that domain, then they
disconnect from the domain. However, each worker thread does have an
associated managed thread which is part of a domain. When the domain
where that managed thread lives is unloaded, the managed thread has its
ThreadState_AbortRequested flag set, and the domain unload waits for the
thread to exit.

This works well as long as the thread stays in the domain. However, the
async IO worker thread can remove the managed thread from an domain that
is in the process of unloading while that domain is waiting for that
thread to exit. When this happens, the domain no longer sees the thread,
so the domain unload proceeds, and the result is a managed thread
object with the ThreadState_AbortRequested flag set that is still used
by the async IO worker thread pool.

The next time an async IO callback uses that thread, the runtime_invoke
call to the managed method will first check if the thread needs to be
interrupted. That check will see the ThreadState_AbortRequested flag
set, and throw a ThreadAbortedException. This exception causes the
thread to silently exit (via a call to ExitThread(1) on Windows, for
example).

So if something is waiting on an asynchronous response that died with
this thread, then that waiter will never see its response.

This is the root of the problem in these two cases.
AuthenticateAsClient does a handshake with the server. It succeeds the
first time, but the second time entering play mode, the Unity editor
unloads the previous app domain, which triggers this problem. The
handshake process starts with the client executing the managed method
RecordProtocol::InternalSendRecordCallback on an async IO thread. The
client then waits in SslStreamBase::EndRead for the response, but
InternalSendRecordCallback was never executed, so the response will not
occur, and EndRead times out.

There is not a good work around for this issue. The AuthenticateAsClient
implementation is entirely in the class library code, so it is not
possible for a user to use synchronous IO to avoid this issue.

The fix for this issue feels pretty safe, as it causes the runtime to
notice the ThreadState_AbortRequested in the same place that it handles
shutdown. So the async IO worker thread should exit cleanly and
decrement the proper global counters.