Double check attempt at acquiring monitor deal with race condition (case 827984).
There is a race between first attempt to acquire the monitor and incrementing entry_count to notify other threads that we will be waiting on sempahore.
The incremental wait logic (waitms) attempts to solve the same problem, but has the disadvantage of causing a 100ms stall in the race condition case.
I think with this check we can remove all the special timeout logic but let's just make a minimal change for now.