Add one more possibility to get a type by name.
For IL2CPP, we don't implement stack walking now. To help get types, we
give a hint about the assembly in the generated code. We need the Mono
code to use the same code path with that hint that we would get if stack
walking found the same assembly. This changes let's us use that code
path.