Remove the NO-JIT configurations in favor .props
In an attempt to mimize the number of changes to .vcxproj files we
need to merge with upstream, I've been able to use a new .sln file and a
new .props file to build the Mono static libraries needed for il2cpp on
mono.

In addition, change the name of libmonoruntime-bdwgc-il2cpp.lib to
libmonoruntime-il2cpp-bdwgc.lib. This is the easiest naming option based
on the way the msbuild files are organized now. I'll make the
corresponding change in the IL2CPP repo.

Finally, update the build scripts to use mono-il2cpp.sln