Remove PLATFORM_UNITY define checks on Windows
We don't want to modify the non-Unity files, so remove the temporary
PLATFORM_UNITY define checks from the files. In order to allow the build
to work, only build the files for the Unity platform.

Use the Windows gdir implementation temporarily.  The libil2cpp OS API
exposure it not ready yet to support this. When it is, we will switch to
it. This step allows us to move forward with the Unity platform changes
in the rest of eglib now though.