* Making the method pointer in SequencePoint struct a simple pointer instead of
a pointer-to-pointer.  We were using a pointer-to-pointer because the pointers
weren't updated at the time we initialized the sequence points, but now we need the
ability to replace those pointers when a shared generic is called.  Fixed the debugger
initialization sequence to allow for this.
* Improved il2cpp_mono_methods_match.  Allowing generic parameters on the left side
to match any type on the right side.  Now that sequence points are the same for all
generic methods this should work in the context of finding sequence points. Also
checking for is_generic to be false and is_inflated to be true before accessing the
genericMethod->methodDefinition field.