* Adding a simplification to mono_domain_get_assemblies_iter in debugger-agent.c that was previously
suggested but didn't make it into the PR.
* Fixing MOD_KIND_SOURCE_FILE_ONLY event request filter in debugger agent.  Switched
to using the new type->source-files map to greatly improve the speed of looking for
a source file in all of the loaded types, converting string to lower case for
case insensitive comparison, fixed the test to search for the correct path.
* Removing mono_object_box calls from do_invoke_method in debugger agent. Replacing
this with a pointer manipulation in il2cpp_mono_runtime_try_invoke(), the same one
that the il2cpp c-api uses for calls on value types.  This fixes the InvokeVType test.