Enable IPv6 support in the settings files
It looks like the upstream Mono code uses either the class library PAL
or a fallback to a hard-coded value to enable IPv6 support. The fallback
is not compiled in for the Unity code, so the class library code tries a
final fallback to the settings file. In the old Mono this setting was
present, but it does not exist in the new Mono, as it is not needed
upstream.

This change corrects Unity case 941946 by enabling IPv6 support in the
settings file.