Work around an Android C++ compiler bug
The class library code in `GetErrorMessage` exposes a C++ compiler bug
on Android in the r13b NDK. We can't easily update the NDK, so instead
we will work around the bug by changing the class library code. We don't
expect any change in behavior from this change.

This corrects case 986215.

Release notes:

IL2CPP: Work around a C++ compiler bug in the Android NDK that occurs when obtaining the message from a socket exception.