Remove the machine.config check for IPv6 support
We want the .NET 2.0 and the .NET 2.0 Subset profile to work the same
with respect to IPv6. Previously, the .NET 2.0 profile checked the
machine.config file for IPv6 support. On Android, we don't ship a
machine.config file, so IPv6 support was not working on Android with the
.NET 2.0 profile.

This change removes the code which checks the machine.config file, and
instead always asks the `Socket` code if it supports IPv6, which should
give us back the proper answer.

This change corrects Unity case 804510 and case 803576.

Release notes:

Add support for IPv6 on Android with the .NET 2.0 profile.