Remove the machine.config check for IPv6 support
We want the .NET 4.x and the .NET Standard 2.0 profiles to work the same
with respect to IPv6.  Previously, the .NET 4.x profile checked the
machine.config file for IPv6 support.

This change removes the code which checks the machine.config file, and
instead always returns true. Later, we will use the Unity PAL to get the
proper value for each platform. For now, we want consistency.