Assume not having libc means this is not macOS
We know that libc will always be present on macOS, so if it is not
there, we can assume this is not macOS.

More generally, use a try/finally to avoid leaking the allocated buffer
when libc is not found and set `checkOS` properly so we don't try to
check again.

This corrects Unity case 1038918, and in general helps our platform
story int he class libraries a little bit, until we get better platform
detection support at runtime.