Link baselib as a static library
This change makes the Mono build link baselib as a static library on
macOS and Android.

At the moment, it finds the baselib static library in the
external/baselib directory, but in the future we will likely get
baselib in a binary package, and external/baselib will disappear.

This change also removes much of the code from the use-baselib branch
which loads baselib as a dynamic library.