Use loop for suspend thread logic (case 1082246)
Calls GetThreadContext (and perhaps SuspendThread) can fail.
This seems to be due to race condition with thread either
being in protected code or invalid state (exiting). Use
loop of SuspendThread/GetThreadContext/ResumeThread(on failure)
to retry suspension logic.

Move GetThreadContext to suspend step and cache in per thread
CONTEXT. This ensures we can get a CONTEXT for each suspended thread.