Implement NamedPipeClientStream in the unityaot profile (case 1159863)
This change builds the class library code to implement the
`NamedPipeClientStream` class in the unityaot profile on Windows. For
the time being, we will not implement this class for non-Windows
platforms, since that requires the Mono.Posix.dll assembly and a native
library as well.

We may consider adding support on Posix platforms in the future.