Fixing issue with debugger CREATE_BOXED_VALUE command. case 1169306.
Rider was sending a CMD_APPDOMAIN_CREATE_BOXED_VALUE command to the
debugger agent when hovering over certain values in the debugger, which
resulted in a class argument that was not a value type.  The debugger
tried to create an object of this type and unbox it, but asserted
because it is not a value type.  Catching this issue before the unbox
command and returning an error prevents the Editor from crashing, but
the debugger still cannot evaulate the value while hovering.  However,
the values can still be inspected using the locals window.