Fixes for HttpUtility.UrlEncode for Unity aot and jit profiles
For aot, this commit moves the HttpUtility back to the System.Web dll. We are keeping a small footprint for System.Web, only what is needed for this class. 

Also the netstandard facade has been updated to include System.Web.

For the Jit profile a change is needed in GetCustomEncoderFromConfig for Android. The jit version of System.Web does not use the MOBILE define as the jit profile is typically used for desktop? This check gives Android a safe fallback.

(case 1305211)`