Fixes for HttpUtility.UrlEncode for Unity aot and jit profiles
Reverting HttpUtility.UrlEncode back to the System.dll to keep
netstandard profiles smaller, especially on Android. In Unity
integration testing we found that adding System.Web to netstandard
profile brings in 4 dlls to Android APKs that were not there before.

This attempt is to leave HttpUtility in System.dll and use a type
forward in System.Web to satisfy .net 4 profile references.