Use only eight characters from the git revision for the artifact name
It looks like the artifactid.txt file that is generated by the job to
upload Mono artifacts to Stevedore is now using only right characters
from the git revision instead of 11. It is unclear why this has changed,
but the number of characters is not material, so long as this value
matches.

This allows the proper artifact name to be generated, so that consumers
of this artifact can reference it correctly.