Fall back to the brokered file system in IL2CPP on Windows (UUM-17611)
For Universal Windows Platform, many kernel32 file system APIs are not
available. In the IL2CPP runtime code, we can fallback to the brokered
filesystem, which uses COM to access these APIs on UWP.

This change requires an associated change in the IL2CPP repository to
implement these new icalls.