This adds a mono call (which unity/burst) will leverage to fake a domain reload on the burst debug domain.
Originally the burst debug domain was torn down and recreated everytime
a bursted library containing metadata was loaded into Unity.
Unfortunately on large projects this can introduce significant stalls,
however we don't actually need to tear down and recreate the domain, we
just need to tell the client that the special type in the burst debug
domain needs to be refreshed.

Note this requires an additional editor change to switch over to the new
functionality, until then, it will still behave as before.